refactor data connector
