+++ sgl-router/src/data_connector/mod.rs
@@ -1,36 +1,25 @@
 // Data connector module for response storage and conversation storage
-pub mod conversation_item_memory_store;
-pub mod conversation_item_oracle_store;
-pub mod conversation_items;
-pub mod conversation_memory_store;
-pub mod conversation_noop_store;
-pub mod conversation_oracle_store;
-pub mod conversations;
-pub mod response_memory_store;
-pub mod response_noop_store;
-pub mod response_oracle_store;
-pub mod responses;
+//
+// Simplified module structure:
+// - core.rs: All traits, data types, and errors
+// - memory.rs: All in-memory storage implementations
+// - noop.rs: All no-op storage implementations
+// - oracle.rs: All Oracle ATP storage implementations
+// - factory.rs: Storage creation function
 
-pub use conversation_item_memory_store::MemoryConversationItemStorage;
-pub use conversation_item_oracle_store::OracleConversationItemStorage;
-pub use conversation_items::{
-    ConversationItem, ConversationItemId, ConversationItemStorage, ConversationItemStorageError,
-    ListParams as ConversationItemsListParams, NewConversationItem,
-    Result as ConversationItemsResult, SharedConversationItemStorage,
-    SortOrder as ConversationItemsSortOrder,
-};
-pub use conversation_memory_store::MemoryConversationStorage;
-pub use conversation_noop_store::NoOpConversationStorage;
-pub use conversation_oracle_store::OracleConversationStorage;
-pub use conversations::{
-    Conversation, ConversationId, ConversationMetadata, ConversationStorage,
-    ConversationStorageError, NewConversation, Result as ConversationResult,
-    SharedConversationStorage,
-};
-pub use response_memory_store::MemoryResponseStorage;
-pub use response_noop_store::NoOpResponseStorage;
-pub use response_oracle_store::OracleResponseStorage;
-pub use responses::{
-    ResponseChain, ResponseId, ResponseStorage, ResponseStorageError, SharedResponseStorage,
-    StoredResponse,
-};
+mod core;
+mod factory;
+mod memory;
+mod noop;
+mod oracle;
+
+// Re-export all core types
+pub use core::*;
+
+// Re-export all storage implementations
+pub use memory::*;
+pub use noop::*;
+pub use oracle::*;
+
+// Re-export factory function
+pub use factory::create_storage;
